chatting-app-django/requirements.txt
Django
django-widget-tweaks
djangorestframework
pytz
django-environ
sympy
numba
pyaes